Code edits
I added the data for the choices and abandonment (10 trials) for the ants that were initially not included in the analysis. Reason: The videos were tilted or of bad quality or ants were not visible in the video if they went underneath the setup. But regardless of that, I recorded the choices during the experiment.